refactor(workspace): extract git autosave into a standalone primitive
Git is history/backup, never the sync engine and never part of materialization
or reconcile, so it does not belong threaded through the markdown seam's every
write path via a markDirty callback. Pull it out into attachGitAutosave, a
self-contained primitive that watches a directory (fs.watch, recursive) and
debounce-commits its changes, knowing nothing about Yjs rows or the vault. It
hooks a Y.Doc's destroy for teardown so it disposes with the workspace it
accompanies.

The vault and export seams lose their git? option and markDirty wiring entirely;
the shared materialize/rebuild core no longer takes a markDirty callback. The
three app mounts that exposed a git? option now compose attachGitAutosave next to
the vault when the option is set, instead of passing git into the seam. Net: the
seam is git-unaware, and git autosave is reusable over any directory.
