+++ apps/honeycrisp/project.ts
@@ -11,6 +11,7 @@ import { EPICENTER_API_URL } from '@epicenter/constants/apps';
 import { defineActions, defineWorkspace } from '@epicenter/workspace';
 import { defineMount } from '@epicenter/workspace/daemon';
 import {
+	attachGitAutosave,
 	attachMarkdownVault,
 	type GitAutosaveConfig,
 } from '@epicenter/workspace/document/materializer/markdown';
@@ -63,8 +64,14 @@ export function honeycrisp(opts: HoneycrispMountOptions = {}) {
 			const markdown = attachMarkdownVault(workspace, {
 				dir: mdDir,
 				tables: { notes: {} },
-				git: opts.git,
 			});
+			if (opts.git) {
+				attachGitAutosave({
+					ydoc: workspace.ydoc,
+					dir: mdDir,
+					config: opts.git,
+				});
+			}
 
 			const actions = defineActions({
 				...workspace.actions,
